# business-central-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Business Central Backend API
  version: 0.1.0
  description: |
    Contract-first OpenAPI documentation for the Business Central backend.

    Health, authentication, authorization, commerce, reporting, service, and
    repair endpoints are implemented. Reusable schemas, security conventions,
    response envelopes, required error codes, and collection query semantics
    are documented here.
  license:
    name: Proprietary
  x-business-central-status: implemented-current-scope
servers:
  - url: http://localhost:8080
    description: Local backend (when the HTTP runtime is available)
...
paths:
...
  /api/v1/inventory/receivable-lines:
    get: { tags: [Inventory], summary: List issued purchase-order lines with quantities remaining, operationId: listReceivableLines, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: Purchase-order lines available for stock receipt, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
  /api/v1/inventory/storage:
    get:
      tags: [Inventory]
      summary: List the view-only stock storage table
      operationId: listInventoryStorage
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Query' }
        - { $ref: '#/components/parameters/Filter' }
        - { $ref: '#/components/parameters/PageIndex' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: sort, in: query, required: false, schema: { type: string, example: stock_count:desc } }
      responses:
        '200': { description: 'Tenant-scoped stock rows, restricted to the assigned shop for staff. RETAIL sell price is available to merchant and staff users; original price and profit are returned only for merchant-role users', content: { application/json: { schema: { $ref: '#/components/schemas/StorageCollectionEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/inventory/movements:
    get: { tags: [Inventory], summary: List merchant stock movement history, operationId: listInventoryMovements, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: Inventory movement history, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
  /api/v1/inventory/movements/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Inventory]
      summary: Get complete stock movement detail
      operationId: getInventoryMovementDetail
      description: Returns the tenant- and shop-scoped immutable movement with resolved product, variant, unit, locations, current balances, linked receipt/purchase order or checkout order, and FIFO cost allocations.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Complete stock movement detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockMovementDetailEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/transaction-history:
    get:
      tags: [Operations]
      summary: List all operational transactions for the selected shop
      operationId: listTransactionHistory
      description: Combines stock-in, stock-out, returns, adjustments, canonical orders from every channel, successful refunds, and repair deposit/final checkout payments. This is intentionally separate from the stock movement ledger and financial report aggregates.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Shop-scoped operational transaction history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionHistoryCollectionEnvelope' }
  /api/v1/transaction-history/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Operations]
      summary: Get complete transaction and canonical order detail
      operationId: getTransactionHistoryDetail
      description: Resolves an accessible history event to its canonical order and returns line item identity, quantity, selling price, FIFO cost, profit, totals, payments, and refunds. Stock events without an order return the event with empty order collections.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Tenant- and shop-scoped transaction detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionHistoryDetailEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/inventory/stock-in:
    post: { tags: [Inventory], summary: Receive stock into a location, operationId: stockIn, security: [{ bearerAuth: [] }], description: "Creates a canonical receipt and FIFO cost layer. When unit_cost is omitted, the latest receipt cost for the merchant-owned variant is reused; the first receipt requires an explicit cost. For direct stock-in, omit purchase_order_id, purchase_order_line_id, and receipt_number. When receiving a purchase order, both IDs and receipt_number are required.", requestBody: { required: true, content: { application/json: { schema: { type: object, required: [variant_id, destination_location_id, quantity, event_key], properties: { purchase_order_id: { type: string, format: uuid }, purchase_order_line_id: { type: string, format: uuid }, variant_id: { type: string, format: uuid }, destination_location_id: { type: string, format: uuid }, unit_id: { type: string, format: uuid }, receipt_number: { type: string }, batch_number: { type: string }, expires_at: { type: string, format: date-time }, quantity: { type: string }, unit_cost: { type: string, description: "Optional after the first receipt; defaults to the variant's latest receipt cost." }, event_key: { type: string } } } } } }, responses: { '201': { description: Stock received and FIFO cost layer created, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/inventory/stock-checkout:
    post: { tags: [Inventory], summary: Check stock out for a sale, operationId: stockCheckout, security: [{ bearerAuth: [] }], description: Requires an existing order_line_id., requestBody: { required: true, content: { application/json: { schema: { type: object, required: [order_line_id, variant_id, source_location_id, quantity, event_key], properties: { order_line_id: { type: string, format: uuid }, variant_id: { type: string, format: uuid }, source_location_id: { type: string, format: uuid }, quantity: { type: string }, event_key: { type: string } } } } } }, responses: { '201': { description: Stock checked out, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/pricing/price-lists:
    get: { tags: [Catalog], summary: List merchant price lists, operationId: listPriceLists, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: Price lists, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
    post: { tags: [Catalog], summary: Create a price list, operationId: createPriceList, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [code, currency_code], properties: { code: { type: string }, currency_code: { type: string }, is_default: { type: boolean } } } } } }, responses: { '201': { description: Price list created, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/pricing/price-lists/{id}/prices:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get: { tags: [Catalog], summary: List prices in a price list, operationId: listPrices, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: Product prices, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
  /api/v1/pricing/prices:
    post: { tags: [Catalog], summary: Create or replace a product price, operationId: upsertPrice, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [price_list_id, variant_id, amount], properties: { price_list_id: { type: string, format: uuid }, variant_id: { type: string, format: uuid }, amount: { type: string }, valid_from: { type: string, format: date-time }, valid_until: { type: string, format: date-time } } } } } }, responses: { '200': { description: Product price saved, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/promotions:
    get: { tags: [Catalog], summary: List merchant promotions, operationId: listPromotions, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: Promotions, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
    post: { tags: [Catalog], summary: Create a promotion, operationId: createPromotion, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, promotion_type, value], properties: { name: { type: string }, promotion_type: { type: string, enum: [PERCENTAGE, FIXED_AMOUNT] }, value: { type: string }, minimum_subtotal: { type: string }, usage_limit: { type: integer } } } } } }, responses: { '201': { description: Promotion created, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/promotions/codes:
    post: { tags: [Catalog], summary: Create a promotion code, operationId: createPromotionCode, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [promotion_id, code], properties: { promotion_id: { type: string, format: uuid }, code: { type: string }, usage_limit: { type: integer } } } } } }, responses: { '201': { description: Promotion code created, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/promotions/{id}/codes:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get: { tags: [Catalog], summary: List codes for a promotion, operationId: listPromotionCodes, security: [{ bearerAuth: [] }], responses: { '200': { description: Promotion codes, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
  /api/v1/promotions/{id}/products:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get: { tags: [Catalog], summary: List product and variant scopes for a promotion, operationId: listPromotionProducts, security: [{ bearerAuth: [] }], responses: { '200': { description: Promotion product scopes, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send the access token returned by login or refresh as
        Authorization: Bearer <access-token>. Refresh tokens are rotated by
        POST /api/v1/auth/refresh.
  parameters:
...
    ResourceID:
      name: id
      in: path
      required: true
      description: Merchant resource UUID.
      schema:
        type: string
        format: uuid
...
    Query:
      name: query
      in: query
      required: false
      description: Case-insensitive text search over the endpoint's searchable fields.
      schema: { type: string }
    Filter:
      name: filter
      in: query
      required: false
      description: Comma- or semicolon-separated equality filters, for example status:OPEN,is_active:true.
      schema: { type: string }
    PageIndex:
      name: page_index
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
...
    Unauthorized:
      description: Authentication is required or the credentials are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist in the authorized tenant context.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    StorageCollectionEnvelope:
      allOf:
        - $ref: '#/components/schemas/CollectionEnvelope'
...
    UUID:
      type: string
      format: uuid
      description: UUID identifier.
...
    ResourceMeta:
      type: object
      description: Optional metadata for a single-resource response.
      additionalProperties: true
    CollectionMeta:
      type: object
      required:
        - page
        - page_size
        - total
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
      additionalProperties: false
    ResourceEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: object
          description: Concrete endpoint resource payload.
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/ResourceMeta'
      additionalProperties: false
    CollectionEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          description: Concrete endpoint resource collection.
          items:
            type: object
            additionalProperties: true
        meta:
          $ref: '#/components/schemas/CollectionMeta'
      additionalProperties: false
...
    TransactionHistoryCollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/TransactionHistoryEntry' } }
        meta: { $ref: '#/components/schemas/CollectionMeta' }
...
    TransactionHistoryDetailEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/TransactionHistoryDetail' }
        meta: { type: object }
...
    StockMovementDetailEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/StockMovementDetail' }
        meta: { type: object }
...
    ErrorResponse:
      type: object
      required:
        - error
        - request_id
      properties:
        error:
          $ref: '#/components/schemas/Error'
        request_id:
          $ref: '#/components/schemas/UUID'
      additionalProperties: false
...